requirements.txt
# keep in alphabetical order
dm_control
interbotix_xs_modules
interbotix_xs_msgs
numpy-quaternion
Pillow
pygame
pyspacemouse
PyQt6
pyquaternion
pyrealsense2
pure-python-adb
quaternion
termcolor
tyro
ur-rtde
xarm
xarm-python-sdk
zmq
